Reorganize components folder for MainDashboard
