+++ src/components/MainDashboard/FilterTopbar.js
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from 'react';
-import {useStateValue} from '../hooks/useStateValue';
+import {useStateValue} from '../../hooks/useStateValue';
 import {Card, Form, Row, Input, Icon, Divider} from 'antd';
-import {StyledCheckableTag as CheckableTag} from '../styled';
+import {StyledCheckableTag as CheckableTag} from '../../styled';
 import OrganizationsList from './OrganizationsList';
 
 const tabList = [
