+++ src/components/MainDashboard/OrganizationCard.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {StyledCard, StyledButton} from '../styled';
+import {StyledCard, StyledButton} from '../../styled';
 import {Link} from 'react-router-dom';
 
 const OrganizationCard = ({org, history}) => {
