+++ src/components/MainDashboard/index.js
@@ -1,0 +1,6 @@
+export * from './FilterTopbar'
+export * from './FilteredComponent'
+export * from './NoEventsFound'
+export * from './EventList'
+export * from './OrganizationCard'
+export * from './OrganizationsList'
