+++ src/components/index.js
@@ -1,12 +1,12 @@
 export * from './Event';
-export * from './EventList';
+export * from './MainDashboard/EventList';
 export * from './Navigation';
 export * from '../styled/StyledUploadImage';
 export * from './FooterDiv';
 export * from './HeaderDiv';
 export * from './UploadImage';
-export * from './FilteredComponent';
-export * from './FilterTopbar';
-export * from './NoEventsFound';
+export * from './MainDashboard/FilteredComponent';
+export * from './MainDashboard/FilterTopbar';
+export * from './MainDashboard/NoEventsFound';
 export * from './ProgressBar';
-export * from './StyledLoader';
+export * from '../styled/StyledLoader';
