+++ src/styled/StyledLoader.js
